move validation to kvdbserver and refactor it
